ft/strmapi: change data type from 'size_t' to 'unsigned int' for consistency

+++ ft_strmapi.c
@@ -6,7 +6,7 @@
 /*   By: frodrig2 <frodrig2@students.42porto.com>   +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/05/21 01:47:09 by frodrig2          #+#    #+#             */
-/*   Updated: 2026/05/21 02:22:45 by frodrig2         ###   ########.fr       */
+/*   Updated: 2026/05/21 02:38:10 by frodrig2         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 char *ft_strmapi(char const *s, char (*f)(unsigned int, char))
 {
-	char	*s2;
-	size_t	i;
+	char			*s2;
+	unsigned int	i;
 
 	if (!s || !f)
 		return (NULL);
